move ex09 renderer to its own include
